refactor: test 코드 수정 및 개선

+++ backend/sin/src/main/java/sin/sin/domain/product/Product.java
@@ -50,7 +50,7 @@ public class Product {
     @JoinColumn(nullable = false, name = "product_category_id")
     private ProductCategory productCategory;
 
-    private float discountPercent;
+    private int discountPercent;
 
     private String saleUnit;
